Move IsCheckoutOpcode to shared package
