+++ src/vm/shared/is_checkout_opcode.go
@@ -1,11 +1,10 @@
-package program
+package shared
 
 import (
 	"github.com/git-town/git-town/v9/src/gohacks"
-	"github.com/git-town/git-town/v9/src/vm/shared"
 )
 
-func IsCheckoutOpcode(opcode shared.Opcode) bool {
+func IsCheckoutOpcode(opcode Opcode) bool {
 	typeName := gohacks.TypeName(opcode)
 	return typeName == "Checkout" || typeName == "CheckoutIfExists"
 }
