+++ src/vm/shared/is_checkout_opcode_test.go
@@ -1,11 +1,11 @@
-package program_test
+package shared_test
 
 import (
 	"testing"
 
 	"github.com/git-town/git-town/v9/src/domain"
 	"github.com/git-town/git-town/v9/src/vm/opcode"
-	"github.com/git-town/git-town/v9/src/vm/program"
+	"github.com/git-town/git-town/v9/src/vm/shared"
 	"github.com/shoenig/test/must"
 )
 
@@ -15,18 +15,18 @@ func TestIsCheckout(t *testing.T) {
 	t.Run("given an opcode.Checkout", func(t *testing.T) {
 		t.Parallel()
 		give := &opcode.Checkout{Branch: domain.NewLocalBranchName("branch")}
-		must.True(t, program.IsCheckoutOpcode(give))
+		must.True(t, shared.IsCheckoutOpcode(give))
 	})
 
 	t.Run("given an opcode.CheckoutIfExists", func(t *testing.T) {
 		t.Parallel()
 		give := &opcode.CheckoutIfExists{Branch: domain.NewLocalBranchName("branch")}
-		must.True(t, program.IsCheckoutOpcode(give))
+		must.True(t, shared.IsCheckoutOpcode(give))
 	})
 
 	t.Run("given another opcode", func(t *testing.T) {
 		t.Parallel()
 		give := &opcode.AbortMerge{}
-		must.False(t, program.IsCheckoutOpcode(give))
+		must.False(t, shared.IsCheckoutOpcode(give))
 	})
 }
